+++ src/resource/level_resource.cpp
@@ -228,7 +228,7 @@ namespace level_resource
 
 	void* load(File& file, Allocator& a)
 	{
-		const size_t file_size = file.size();
+		const uint32_t file_size = file.size();
 		void* res = a.allocate(file_size);
 		file.read(res, file_size);
 		return res;
